[core] Expose a library entry point
Command parsing, process setup, dispatch, and every production
module lived under the binary target. A test driver or alternate
composition root could only reuse that behavior by executing the
shipping binary.

Move the existing implementation unchanged into the library and
reduce the binary to one call. This creates a reusable production
boundary without changing behavior or injecting dependencies yet.

Keep the command and regex helper macros crate-private. They were
internal implementation details in the binary and should not become
unsupported library API merely because the modules moved. Mark the
command macro usage block as illustrative text so Rustdoc does not
present it as a downstream API.

gherrit-pr-id: Gz2tpe6hciedj2oun55jv42qojb4lsthn
